requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
ollama==0.4.0
faster-whisper>=1.1.0
beautifulsoup4==4.12.0
httpx==0.27.0
duckduckgo-search==6.0.0
pydantic-settings==2.0.0
aiosqlite==0.20.0
chromadb>=0.5.5
python-multipart==0.0.9
pymupdf==1.24.10
python-docx==1.1.0
feedparser==6.0.11
